Issue #1356: Fix 'Output this field as link' in views.
In a view, the 'Output this field as link' option wasn't rewriting the
tokens from 'Link path'. The old code was only calling
`get_render_tokens` if `$value` contains a '[', but the tokens to be
replaced aren't in `$value`, they're in `$alter['path']`, so `$tokens`
was just an empty array, which is why none of the tokens were being
replaced.